vector_db/requirements.txt
fastapi==0.110.0
langchain==0.1.11
pandas==1.5.3
pinecone-client==3.1.0
pydantic==2.6.3
python-dotenv==1.0.1
tqdm==4.66.2
sentence_transformers==2.2.2
uvicorn==0.27.1
